Change batteryLevel type from int to float

+++ Sources/Client/MBTClientV2.swift
@@ -113,7 +113,7 @@ public class MBTClientV2 {
     return deviceInformation?.productName
   }
 
-  public var lastBatteryLevel: Int {
+  public var lastBatteryLevel: Float {
     return bluetoothManager.lastBatteryLevel
   }
 
@@ -554,7 +554,7 @@ public class MBTClientV2 {
 
 extension MBTClientV2: MBTBluetoothAcquisitionDelegate {
 
-  public func didUpdateBatteryLevel(_ levelBattery: Int) {
+  public func didUpdateBatteryLevel(_ levelBattery: Float) {
     acquisitionDelegate?.didUpdateBatteryLevel(levelBattery)
   }
